backend/pyproject.toml
[project]
name = "back"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi>=23.1.0",
    "django>=5.2",
    "dotenv>=0.9.9",
    "psycopg2-binary>=2.9.10",
]
